epi-83: removed redundant files
